fixes after feedback
